Fix issues while unit testing and add missed endpoints

+++ Backend/Business/Flinnt.Interfaces/Services/IPostUserService.cs
@@ -8,9 +8,9 @@ namespace Flinnt.Interfaces.Services
     public interface IPostUserService : IBaseService<PostUser>
     {
         Task<List<PostUserViewModel>> GetAllAsync(int postId);
-        Task<PostUserViewModel> GetAsync(int id);
+        Task<PostUserViewModel> GetAsync(long id);
         Task<bool> AddAsync(PostUserViewModel model);
         Task<bool> UpdateAsync(PostUserViewModel model);
-        Task<bool> DeleteAsync(int id);
+        Task<bool> DeleteAsync(long id);
     }
 }
